* Improve memory operation performance * Fix division bug by switching to rationals * Add division regression test

+++ borealis/borealis/src/rudder/opt/constant_folding.rs
@@ -225,7 +225,7 @@ fn cast_integer(value: ConstantValue, typ: Arc<Type>) -> ConstantValue {
                 ConstantValue::UnsignedInteger(i)
             }
             (PrimitiveTypeClass::FloatingPoint, ConstantValue::SignedInteger(s)) => {
-                ConstantValue::FloatingPoint(f32::try_from(i16::try_from(s).unwrap()).unwrap())
+                ConstantValue::FloatingPoint(f64::try_from(i16::try_from(s).unwrap()).unwrap())
             }
             (tc, cv) => {
                 panic!("incompatible type class {tc:?} and constant value {cv:?}")
